Divided classes to separate files + cross imports
