+++ App.py
@@ -1,0 +1,54 @@
+import json
+import matplotlib
+
+import numexpr as ne
+import numpy as np
+
+from functools import partial
+from tkinter import *
+from tkinter.filedialog import asksaveasfile
+
+from matplotlib import pyplot as plt
+
+from matplotlib.backends.backend_tkagg import (FigureCanvasTkAgg,
+                                               NavigationToolbar2Tk)
+
+from Buttons import *
+from Commands import *
+from Entries import *
+from ModalWindow import *
+from Plotter import *
+
+# app class (класс приложения)
+class App(Tk):
+    def __init__(self, buttons, plotter, commands, entries):
+        super().__init__()
+        self.buttons = buttons
+        self.plotter = plotter
+        self.commands = commands
+        self.entries = entries
+        self.entries.set_parent_window(self)
+        self.plotter.set_parent_window(self)
+        self.commands.set_parent_window(self)
+        self.buttons.set_parent_window(self)
+
+    def add_button(self, name, text, command_name, *args, **kwargs):
+        hot_key = kwargs.get('hot_key')
+        if hot_key:
+            kwargs.pop('hot_key')
+        callback = partial(self.commands.get_command_by_name(command_name), *args, **kwargs)
+        new_button = self.buttons.add_button(name=name, text=text, command=callback)
+        if hot_key:
+            self.bind(hot_key, callback)
+        new_button.pack(fill=BOTH)
+
+    def get_button_by_name(self, name):
+        return self.buttons.buttons.get(name)
+
+    def create_menu(self):
+        menu = Menu(self)
+        self.config(menu=menu)
+
+        file_menu = Menu(menu)
+        file_menu.add_command(label="Save as...", command=self.commands.get_command_by_name('save_as'))
+        menu.add_cascade(label="File", menu=file_menu)
