+++ Buttons.py
@@ -1,0 +1,39 @@
+import json
+import matplotlib
+
+import numexpr as ne
+import numpy as np
+
+from functools import partial
+from tkinter import *
+from tkinter.filedialog import asksaveasfile
+
+from matplotlib import pyplot as plt
+
+from matplotlib.backends.backend_tkagg import (FigureCanvasTkAgg,
+                                               NavigationToolbar2Tk)
+
+from App import *
+from Commands import *
+from Entries import *
+from ModalWindow import *
+from Plotter import *
+
+# class for buttons storage (класс для хранения кнопок)
+class Buttons:
+    def __init__(self):
+        self.buttons = {}
+        self.parent_window = None
+
+    def set_parent_window(self, parent_window):
+        self.parent_window = parent_window
+
+    def add_button(self, name, text, command):
+        new_button = Button(master=self.parent_window, text=text, command=command)
+        self.buttons[name] = new_button
+        return new_button
+
+    def delete_button(self, name):
+        button = self.buttons.get(name)
+        if button:
+            button.pack_forget()
