+++ Commands.py
@@ -1,0 +1,104 @@
+import json
+import matplotlib
+
+import numexpr as ne
+import numpy as np
+
+from functools import partial
+from tkinter import *
+from tkinter.filedialog import asksaveasfile
+
+from matplotlib import pyplot as plt
+
+from matplotlib.backends.backend_tkagg import (FigureCanvasTkAgg,
+                                               NavigationToolbar2Tk)
+
+from App import *
+from Buttons import *
+from Entries import *
+from ModalWindow import *
+from Plotter import *
+
+# class for commands storage (класс для хранения команд)
+class Commands:
+    class State:
+        def __init__(self):
+            self.list_of_function = []
+
+        def save_state(self):
+            tmp_dict = {'list_of_function': self.list_of_function}
+            file_out = asksaveasfile(defaultextension=".json")
+            if file_out is not None:
+                json.dump(tmp_dict, file_out)
+            return self
+
+        def reset_state(self):
+            self.list_of_function = []
+
+    def __init__(self):
+        self.command_dict = {}
+        self.__figure_canvas = None
+        self.__navigation_toolbar = None
+        self._state = Commands.State()
+        self.__empty_entry_counter = 0
+        self.parent_window = None
+
+    def set_parent_window(self, parent_window):
+        self.parent_window = parent_window
+
+    def add_command(self, name, command):
+        self.command_dict[name] = command
+
+    def get_command_by_name(self, command_name):
+        return self.command_dict[command_name]
+
+    def __forget_canvas(self):
+        if self.__figure_canvas is not None:
+            self.__figure_canvas.get_tk_widget().pack_forget()
+
+    def __forget_navigation(self):
+        if self.__navigation_toolbar is not None:
+            self.__navigation_toolbar.pack_forget()
+
+    def plot(self, *args, **kwargs):
+        def is_not_blank(s):
+            return bool(s and not s.isspace())
+
+        self._state.reset_state()
+        list_of_function = []
+        for entry in self.parent_window.entries.entries_list:
+            get_func_str = entry.get()
+            self._state.list_of_function.append(get_func_str)
+            if is_not_blank(get_func_str):
+                list_of_function.append(get_func_str)
+            else:
+                if self.__empty_entry_counter == 0:
+                    mw = ModalWindow(self.parent_window, title='Пустая строка', labeltext='Это пример модального окна, '
+                                                                                          'возникающий, если ты ввел '
+                                                                                          'пустую '
+                                                                                          'строку. С этим ничего '
+                                                                                          'делать не нужно. '
+                                                                                          'Просто нажми OK :)')
+                    ok_button = Button(master=mw.top, text='OK', command=mw.cancel)
+                    mw.add_button(ok_button)
+                    self.__empty_entry_counter = 1
+        self.__empty_entry_counter = 0
+        figure = self.parent_window.plotter.plot(list_of_function)
+        self._state.figure = figure
+        self.__forget_canvas()
+        self.__figure_canvas = FigureCanvasTkAgg(figure, self.parent_window)
+        self.__forget_navigation()
+        self.__navigation_toolbar = NavigationToolbar2Tk(self.__figure_canvas, self.parent_window)
+        self.__figure_canvas.get_tk_widget().pack(side=TOP, fill=BOTH, expand=1)
+        plot_button = self.parent_window.get_button_by_name('plot')
+        if plot_button:
+            plot_button.pack_forget()
+
+    def add_func(self, *args, **kwargs):
+        self.__forget_canvas()
+        self.__forget_navigation()
+        self.parent_window.entries.add_entry()
+
+    def save_as(self):
+        self._state.save_state()
+        return self
